Data Binding within Fragment
So on first was performing a `getBinding` on the inflater view, however there was no associated binding with the view, and was assigning a value of null to the class member `mBinding`.

The `DataBinding.inflate` method has been used instead to perform the inflating and to get the desired binding.