Baking Adapter Modification for Data
So updating the Baking Application Addapter that I plan on using for the recylcer views.  It is a generic such that the data sets can change for each type of adapter.  It will be required that the views that the adapter works with will likely also need to be templated to be the same as well.

Went ahead and updated a way to update the adapter's data and provided definitino for the count of objects.