AppAdapter Interface Defined
Within the RecipesFragment class, the definition for the AppAdapters interface definition has been completed.
This means that upon the creation of the adapter, the `this` instance of the RecipesFragment class can be provided.

Within the adapter the listener is set during the constructor only.
The `onCreateViewHolder` is responsible for passing the listener to the viewholder upon construction.

The AppView class provides a default action for the `onClick` that calls the AppAdapter defined listener method `onClick`.