ViewHolder Factory
Addition of a `ViewHolder` factory that is responsible as a single point of evaluating what type of `ViewHolders` need to be generated for the `AppAdapter` being using within the Baking Application.

Due to this being a single point and using templated types, needed to know the specific `ViewHolder` that is to actually be generated.  This is now done by augmenting the constructor for the `AppAdapter` to include the specific `ViewHolder` id which is then used in the factory to call the correct constructor.