RecipeStepActivity
Removed several un-used files from the project & cleanup up a few file names.

New activity that own the details w.r.t the recipe step that the user would have clicked or navigated to.  The first fragment that is owned by this activity is the `fragment_recipe_step_detail.description` and its accompanying logic class of `RecipeStepDetailDescriptionFragment.java`.

Since this fragment for the time being only has a text view, went ahead and set the text of the fragment to that of the description as desired within the `onCreateView` of the fragment.

The other fragments have yet to be setup or added to the fragment manager within the `RecipeStepActivity`