BakingActivity ownership of Retrofit Callback
Originally the RecipesFragment was responsible for the owner ship of the retrofit2 callback to obtain the JSON information from teh network.  However, this is coupling logic that the `RecipesFragment` has no need to control.  Instead `BakingActivity` is now the owner of the callback definitions and will perform a replace of the `RecipesFragment` when the asynchronous response returned is valid.