Navigation via OnClickListeners
The RecipeDetailsActivity now defines the `onNavigationClick method for its owned `RecipeStepDetailFragment`.

Based on the current recipe step, will determine what happens with respect to the user interface.  value of -1 will be to display the ingredients.  < -1 or >= the number of steps means to go back to the Recipe Activity via a `finished` call.  Otherwise the next step is displayed on the UI.