Recipe Ingredient Fragment Display
Displaying the list of fragments within the `RecipeDetailsActivity` via the navigation buttons (from steps). The activity can default to the recipe ingredients if the correct index is provided to the intent.

Add the setting of the arguments for the RecipeIngredientFragments so that the `RecyclerView` widget can be populated with data.