Widget User Interface Layout
Want the widget to support both an image and a text field contained within a linear layout.  The plan is that the widget will have an image associated with it and a text field as well. This means that if no image is provided a default one will be used but the name of the recipe should still be contained within the textview.