Widget Stopping Point
At the moment the widget development feature allows the creation of a widget.  However the feature is incomplete so far.  The widget needs to be a list of ingredients, thus the widget process will need to be able to obtain this list.  Need a feature to be able to persist data across process space.