[FTR/BUG] Removal of Room
Feature Develop & Address of BUG

BUG
Current feature development of issue #24 uses a shared component.  Originally was
going to implement persistent data via Room since it is an ORM, however it is not
playing well with the DataBinding Android Architecture component at this time.

Feature
Instead of Room the Baking application will use shared components to provide an
interface for store and inserting persistent data.

Deletion and update are wholly not supported at this time.
Individual ingredient insertion and query are also not supported at this time.

Bug and Feature supports the following issues.
- #24