[FTR] Initial Insert Skeleton LoaderManager.LoaderCallbacks
Feature Development

The initial logic for inserting the recipe's ingredient list has been started.
There is a limitation however that upon every restart of the activity the
network provided data will keep populating the database.

There are a few logical pieces for this to prevent this.  Could specify insertion
rules to be most recent, only if doesnt exist.  However this will still be a large
number of operations potentially on the database.

Only want to insert if the recipe does not already exist.  Thus prior to performing
a network fetch of recipes, read from the database.

This will require that the interaction with the database occurs in the background
from the main thread.

Therefore the skeleton for the LoaderManager to associate a AsyncTask with the main
Activity has been started.

Related Issues
- #26