[FTR] SharedPreference Update
Feature Development

The desired behavior is that when a user is navigating the steps of a recipe, the current
step is recorded in the shared preferences.

If the user underflows or overflows the available set of steps then the shared preferences
will be cleared for the application, and then return back to the `RecipeActivity` view.

Related Issues
- #28