[FTR] Widget Service for Populating ListView Widget
Feature Development

In the context of the ListView widget, the goal is to populate a scrollable listview widget on
the widget for the user.  This is due to the fact that there could be a lot of ingredients for
a given recipe, so a static text view will not be adequate.

At the moment there are two services for the widget.
1. Service to obtain persistent data and pass to the remote views to populate the UI.
2. Service to fill in the ListItem with the actual ingredient information.

Related Issues
- #28