[FTR] Recipe Selection Widget Update
Feature Description

Starting a new recipe will now change the widget list of ingredients and recipe name accordingly.
Have not considered the completion or backing out of a recipe yet.

This requires using the onDataChanged via a receive of a Boardcast method that then performs the
notifyDataChanged state on the specific dynamic widget.

Related Issues
- #28
- #29