[FTR] Clickable ListView and TextView on Widget
Feature Development

Want to be able to support the user clicking on the widget to bring up the application.  This
required some forming from the [reference](https://www.sitepoint.com/killer-way-to-show-a-list-of-items-in-android-collection-widget/).  Would not have been able to figure it out from the ListView widget on click
behavior without it.

Related Issues
- #28
- #29