[FTR] ExoPlayer Video Support
Feature Development

In order to support the possibility of playing videos, a basic setup of the Exoplayer has been
created to support watching.  At the moment the exoplayer should also support the media player
to still support other audio events (calls, alarms) that occur while the exoplayer is playing.

This is supported by making the exoplayer instance be an overall client to the media session
in order to preserve overall state.

It should be noted that a large amount of this follows guidelines from the Udacity course on
`Media` Playback, such that there are some deprecaited elements with respect to the
`onPlayerStateChanged`.

Related Issues
- #32
- #33