[FTR] Exo-Player XML Modifications
Feature Development

Previously the exo-player works but the UI for the exoplayer was not showing
well.  This commit goes into to modifications into layout `xml` files to
account for this.

Previous use of a `ScrollView` widget led to smashed layout within the
recipe details layout.  This has now been removed and getting better
layout.

Related Issues
- #32
- #33