[FTR] Large Device & UI Consolidation
Feature Development

Next feature development has to deal with the UI for large devices.  First phase is
adding in additional layout resources to replace the phone resources.  The beginning
of consolidation of the `RecipeActivity` has also started since with the `Master
View` will need to be able to load in the `RecipeActivityDetail` as well within a
fragment.

Related Issues
- #18
- #30
- #35