[FTR] MasterView for Larger Devices
Feature Development

With the previous work of setting up some device size xml loading, in the `RecipeActivity`
now can operate on this boolean value that indicates if current device is a tablet or a smaller
device due to the setting of the flag.

This flag is used to load in the larger view area for the detail and essentially disables the
`RecipeDetailsActivity` entirely.  This activity will only be loaded on a device with a width
smaller than 600dp.

The new layout for the larger device has a `FrameLayout` widget that is to be populated by the
FragmentManager.  On click behavior has been modified to switch out the fragments to their
appropriate value when navigation has occurred for the larger device.

Related Issues
- #18
- #30
- #35