[FTR] RecipeActivity `onCreate` Update for Multiple Panes
Feature Development

`onCreate` was previously not developed.  `onCreate` now needs to determine the initail
user interface that is based on the value boolean for tablet mode.  If tablet mode is
necessary then the normal array of steps and ingredient selection is presented in pane one
while the detail is populated in pane two.

If the recipe had been previously selected by the user, the the last detail view that was
visited is displayed as well.  Otherwise the ingredient detail fragment is displayed
instead.

Related Issues
- #18
- #30
- #35