[BUG] Orientation Support Large Device
BUG Description

Previous the `flDetails` fragment was not being populated.  Went ahead
and forced a minimum width to isolate the issue to be tablets only.
Also I was using the generic parent `Fragment` class for inserting into
the `FragmentManager` this is however the primary issue that causes the
exception on the devices.

Modified the code that forces separate commit for the different
possibilities for the `flDetails` fragment.

Related Issues
- #39