[FTR] onClick Recipe Espresso Testing
Feature Development

Going through each of the recipes and performing an onclick to bring to
the front the `RecipeActivity` activity.

The title is then compared to the recipe that was clicked and then the
up button is pressed to go back to the `BakingActivity`.  States of the
activity are checked to make sure the correct state transitions are
occurring.