[FTR] Explayer for Thumbnail
Feature Development

Went ahead and performed the same logic that was used for the video
stream for the thumbnails.

This also included some slight changes in logic for the layout such that
the linear layout widgets containing the video and thumbnails will no
longer show at all if they do not exist for the specific step.

Related Issues
- #33