[BUG] Exoplayer Play Loop
BUG Description

When performing a rotation on a small factor device if the detail had an
ExoPlayer widget executing it would keep the old fragment and overlay a
new fragment on top of it (an onCreate call is performed but from an instance
transition that already had an exoplayer resource).

Now the `RecipeDetailsActivity` performs the same logic as a larger device
for controlling the fragment panes, that is it inspects if `onSavedInstance`
bundle is not null.  If null then uses the addition of new fragments to the
fragment manager, otherwise it performs a replace operation which will envoke
the destory method of the previous exoplayer resources.

Related Issues
- #54