[BUG] Tablet and Phone Support Single Exoplayer
BUG Description

On both tablets and phone there was chance of getting duplicate playback
of video streams.  This occurred due to the handling of the fragments in
the logic hosting activity of the fragments.

Once moved to single instance had a reaplce vs add design.  In the
constructor however dont want to perform replace due to wiping fragment
state that may be necessary.  So just support add when the savedinstance
is set to null.  Otherwise the fragment manager will restart the
fragment once onCreate in the host activity is finished.

Related Issues
- #54